Remove comparision tests
The comparison tests have not been part of the test suit in months due
to them being machine dependant.  It also required a long and tedious
interaction between the developer and the outputs, taking upwards
of an hour.  It was also extremely slow, rendering thousands of images.
Nor did it actually reflect a benefit to the game, as these did not
ensure that the editor reflected the game.  For this reason, the
tests have been removed.